fix: (c): double -> float

+++ src/c/kamada_kawai/main.c
@@ -53,7 +53,7 @@ int main(int argc, char *argv[])
     }
 
     int n;
-    double k;
+    float k;
     int **vertices;
     int **edges;
